java-sondaggio-OS completato

+++ src/org/lessons/java/security/PasswordGenerator.java
@@ -9,8 +9,8 @@ public class PasswordGenerator {
 		String fname;
 		String lname;
 		String favColor;
-		byte birthDay;
-		byte birthMonth;
+		int birthDay;
+		int birthMonth;
 		int birthYear;
 		String pwd;
 		
@@ -22,9 +22,9 @@ public class PasswordGenerator {
 		System.out.println("Inserisci il tuo colore preferito:");
 		favColor = input.nextLine();
 		System.out.println("Inserisci il giorno in cui sei nato:");
-		birthDay = input.nextByte();
+		birthDay = input.nextInt();
 		System.out.println("Inserisci il mese in cui sei nato:");
-		birthMonth = input.nextByte();
+		birthMonth = input.nextInt();
 		System.out.println("Inserisci l'anno in cui sei nato:");
 		birthYear = input.nextInt();
 		pwd = fname + "-" + lname + "-" + favColor + "-" + (birthDay+birthMonth+birthYear);
